requirements.txt
streamlit==1.41.1
anthropic==0.42.0
python-docx==1.2.0
python-dotenv==1.2.2
authlib>=1.3.2
